Implemented PaginateWindow. Improved EconomyManager. Added network status to resftul API. Implemented skyblock shop.

+++ API/Economy/src/main/java/pl/north93/zgame/api/economy/ICurrency.java
@@ -4,5 +4,5 @@ public interface ICurrency
 {
     String getName();
 
-    int getStartValue();
+    double getStartValue();
 }
